+++ packages/reporters/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./renderMarkdownReport.js";
+export * from "./renderAgentExecutionPlan.js";
 export * from "./renderHtmlReport.js";
 export * from "./lintReport.js";
 export * from "./writeReport.js";